Add extension to ViewCommand
Add dayToFind and dateToFind variables to
ScheduleContainsKeywordPredicate in order to filter the person list
according to the extension.

The above change allows the application to display the search term the
user actually input.

The above change allows ViewCommand functionality to look for timings
based on the extension's search definition.

Also modified ViewCommandTest to reflect changes.